+++ src/modules/pathfinding/closest-enemy-pathfinding-cron.ts
